Fix requirements of GenerateForecastErrors tests
Modify requirements to apply the same way of generating list of forecast errors like in case of daily errors. It is about the case when there are irregular requests - every other day they are late by the same period. Every day the number of calculated errors during the day is constant.