Move download, save, calc data to hosted service
Thanks to this change, the application download, save and calculate data
on server side automatically once a day. Until now, a customer had to
visit the website at once a day at a certain time to do all these actions.
Additional benefit of the change will be faster page loading.