Apply ICollectionFixture to RepositoryTests
Thanks to this change, setup code for Unit Test classes using the same
objects is in one common place - they have shared context. Additionally
Seed() method for DbContext is changed, because DbContext now is created
once at start of test collection and it is disposed after all tests
belonging  to the test collection. Therefore it is necessary to call
Remove methods on DbContext, because without them entities are tracked
so on (they are still in Local collection of DbSet).