Fix reference to be more suggestive - _installationIds
measurementsList always has as many elements as _installationIds. If for
given installation there wasn't need to download new data, empty Measurements object is added to measurementsList element. This is
because in Measurements class there is no InstallationId property and consecutive Measurements objects correspond to consecutive
installationIds.

It is possible to add InstallationId property to Measurements class
and it would be assigned after downloading. Then measurementsList
would be independent of _installationIds.